feat: 서비스별 입력값 검증 및 파일/채널 삭제 로직 개선
[BinaryContentService, ReadStatusService, UserStatusService]
- delete : 파일 존재 여부 검증 추가

[ChannelService]
- delete : 채널 삭제 시 채널 메시지 관련 파일 삭제 로직 추가

[MessageController]
- 수정/삭제시 사용자 검증을 위해 userId 입력값 추가

[MessageService]
- create: 수신자가 사용자일 경우 존재 여부 검증 및 비공개 채널 접근 검증 추가
- findLastestMessage : 송수신자 존재 여부 확인 검증 로직 추가
- update, delete : 요청한 사용자와 메시지 발신자 일치 여부 검증 추가

[MessageRepository]
- deleteByChannelId : 메시지 관련 파일 삭제를 위해 binaryContent Id 리스트 반환

[ErrorCode]
- MESSAGE_EDIT_FORBIDDEN : 사용자와 메시지 발신자 불일치시 예외 추가

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicBinaryContentService.java
@@ -49,6 +49,9 @@ public class BasicBinaryContentService implements BinaryContentService {
 
     @Override
     public void delete(UUID binaryContentId) {
+        if(!binaryContentRepository.isExist(binaryContentId)){
+            throw new CustomException(ErrorCode.BINARYCONTENT_NOT_FOUND);
+        }
         binaryContentRepository.delete(binaryContentId);
     }
 }
